Semi-working left and right movement
Some testings with a RigidBody2D for the character. It falls with gravity and can be controlled to go left, up, and right, but it seems a bit weird plus it's simpler than it needs to be. More testings are needed to validate this type of node.
Important note : I do not know how it will handle collisions, or if it will at all. This is important because collisions with platforms should trigger the "jump".