Change on respawn, score updated, added HUD, added retry possibility, added exported (Windows) version
The HUD has been added to include the functioning score, messages to tell the player what to do, a working button for restart, which has been reworked to allow the respawn of the player and the de-pop of the platforms upon game over.

The project has been exported to a .exe with accompanying files (and in a compressed file), with the icon being changed.

Next up:
- Visual upgrades, in particular the platforms
- Implement a high score of some sort