Added highscore
Added a highscore display. The highscore is stored in a file in res:// (it is in Appdata/Godot/App_userdata/Name or something). The max between the score and the highscore is displayed and written in the file.

Warning: this has not been tested with the export build, I don't know what happens if it's not used in the usual directory.