Comments + solve score abuse + fix start game message kinda
Some comments have been added to hud.gd.

The abuse on the score consisting on not starting to jump and letting the platforms disappear to earn points has been solved by only starting the point count when the player jumps on a moving platform for the first time.

The message telling the player to press space to start jumping has been improve to last until the player actually does press the key. The solution led to issues because it is ugly and creates lots of useless timers, that might overload the game. This could be solved by reworking the message system, or worked around by deleting this message and have the jumping start from the beginning.