Added platform spark animation
Added an ugly spark animation for when the platforms disappear at the bottom of the screen. This is not very elegant as uses the animation of the platform, requiring the platform to disappear at a precise point (not very scalable). This will need to be improved in the future.

Also started making a new button for the instructions.