Changed platforms + bouncing + font size scaling
Made the platforms thicker and cropped the corners so the Player bouncing on it works better (smoother edges means fairer bounce on the corners.

Changed the way the Player collides with walls and roof so that it now bounces on them. This should allow for better plays using the bounce.

Adapted the size of the instruction label to the size of the window so that it (should) be displayed fully even on smaller screens.