Added starter platforms + changed platform animation
Added platforms at the start, in order to make the game start faster. Added a tweak so that the explosion animation plays correctly (made the platforms stop at the correct location).

Changed the animation played when a platform exits the screen from sparks to an explosion. The entirety of the animation isn't played, because the time is too short, this could be changed in the future.

Tried to improve the jump so that it's not bugged when on edges, but this changes the jump mechanic big time so idk if it's worth it.

Probably other things I forget.