Changed font + added options + changed end button
The font for all the labels and buttons has been changed to Bonus Coffee, a free font from the Internet. This choice can be discussed as it was a bit random.

Added a new button on the main screen that opens an options panel (the aspect is temporary). For now, the only option available is the choice whether the start is a fast one, id est whether the first platforms appear close to the ground or at the top of the screen. The implementation of this option has been realized.

Changed the way the EndButton works, it now saves the score, and displays a Game Over message.