Added moving background
For now it is only an ugly version, but the background has been changed to add moving parts that scroll down (starting when you reach a certain score) (possibly, the choice of the threshold could be given to the player in the options). The original drawing is kept, but in the final version it should be Google Search screenshots instead of ugly rectangles. The code should handle the transition correctly.