Added highscore animation + choice for light or dark moving background + intro animation + change on background animation
Upon reaching a new highscore, the highscore briefly turns gold to celebrate.

The player now has the choice in the options between a light and a dark moving background.

There is now an intro animation played if the player wants (in options) at the beginning of a game.

The moving background now stops and disappear when the game ends (to avoid issues with UI and sprites still present at the beginning of a new game.