Changed ground + added story + commented a bit
Changed the ground sprite to earth with grass so that it is less weird than a black rectangle. The walls and roof changed too...

Added a little bit of a story to the game, for a bit of added depth.

Added some comments to the code. Added credits in the form of a file as well.