Changed explosion animation + changed platform spawn rate + made options default to on
Changed to explosion animation for one that fits correctly into the frame (no weird shapes on the edges).

Changed the way the platforms spawn so that there is more at the beginning and less and the end. Adapted the max speed so that it should remain playable throughout the game.

The options now default to "on".